counting sort and radix sort

+++ utils/limits.go
@@ -4,7 +4,7 @@ const (
 	MIN_INT = 0
 	// use math.*
 	MAX_UINT  = ^uint(0)
-	MIN_UINT  = 0
+	MIN_UINT  = uint(0)
 	MAX_INT   = int(MAX_UINT >> 1)
 	MIN_INT_2 = -MAX_INT - 1
 )
